requirements.txt
pywin32==224
redis==2.10.6
pytest==3.8.2
lxml==3.4.4
requests_toolbelt==0.8.0
kafka==1.3.5
psutil==5.4.8
allure_python_commons==2.5.2
requests==2.18.4
paramiko==2.4.2
urllib3==1.22
python_jenkins==1.2.1
